se agrego validacion de doc para cambio de roles

+++ src/public/js/cart.js
@@ -9,9 +9,35 @@ const addCart = async (idCart, idProduct) => {
     console.log(response.status);
     const data = await response.json();
     if (response.ok) {
-      alert("Producto agregado correctamente");
+      Toastify({
+        text: "Producto agregado al carrito",
+        duration: 3000,
+        destination: "https://github.com/apvarun/toastify-js",
+        newWindow: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background:
+            "linear-gradient(90deg, rgba(3,2,23,1) 0%, rgba(9,23,110,1) 33%, rgba(0,212,255,1) 100%)",
+        },
+        onClick: function () {}, // Callback after click
+      }).showToast();
     } else {
-      alert("Error al agregar el producto");
+      Toastify({
+        text: "Error al agregar el producto al carrito",
+        duration: 3000,
+        destination: "https://github.com/apvarun/toastify-js",
+        newWindow: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background:
+            "linear-gradient(90deg, rgba(0,0,0,1) 0%, rgba(121,9,9,1) 36%, rgba(255,16,0,1) 100%)",
+        },
+        onClick: function () {}, // Callback after click
+      }).showToast();
     }
     return data;
   } catch (error) {
@@ -28,29 +54,77 @@ const deleteProduct = async (idCart, idProduct) => {
       },
     });
     const data = await response.json();
-    alert("Producto eliminado correctamente");
-    location.reload();
+    Toastify({
+      text: "Se elimino el producto del carrito",
+      duration: 3000,
+      destination: "https://github.com/apvarun/toastify-js",
+      newWindow: true,
+      gravity: "top", // `top` or `bottom`
+      position: "right", // `left`, `center` or `right`
+      stopOnFocus: true, // Prevents dismissing of toast on hover
+      style: {
+        background:
+          "linear-gradient(90deg, rgba(3,2,23,1) 0%, rgba(9,23,110,1) 33%, rgba(0,212,255,1) 100%)",
+      },
+      onClick: function () {}, // Callback after click
+    }).showToast();
+    // location.reload();
     return data;
   } catch (error) {
     console.log(error);
   }
 };
 
+
+
 const deleteAllProducts = async (idCart) => {
-  try {
-    const response = await fetch(`/api/cart/${idCart}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
+  const getCart = await fetch(`/api/cart/${idCart}`);
+  const data = await getCart.json();
+  if (data.products.length === 0) {
+    Toastify({
+      text: "Carrito vacio",
+      duration: 3000,
+      destination: "https://github.com/apvarun/toastify-js",
+      newWindow: true,
+      gravity: "top", // `top` or `bottom`
+      position: "right", // `left`, `center` or `right`
+      stopOnFocus: true, // Prevents dismissing of toast on hover
+      style: {
+        background:
+          "linear-gradient(90deg, rgba(0,0,0,1) 0%, rgba(121,9,9,1) 36%, rgba(255,16,0,1) 100%)",
       },
-    });
-    const data = await response.json();
-    alert("Productos eliminados correctamente");
-    location.reload();
-    return data;
-  } catch (error) {
-    console.log(error);
+      onClick: function () {}, // Callback after click
+    }).showToast();
+  } else {
+    try {
+      const response = await fetch(`/api/cart/${idCart}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const data = await response.json();
+      Toastify({
+        text: "Se vacio el carrito",
+        duration: 3000,
+        destination: "https://github.com/apvarun/toastify-js",
+        newWindow: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background:
+            "linear-gradient(90deg, rgba(3,2,23,1) 0%, rgba(9,23,110,1) 33%, rgba(0,212,255,1) 100%)",
+        },
+        onClick: function () {}, // Callback after click
+      }).showToast();
+      
+      return data;
+    } catch (error) {
+      console.log(error);
+    }
   }
+  
 };
 
 const checkOut = async (idCart) => {
@@ -65,6 +139,20 @@ const checkOut = async (idCart) => {
     location.replace(`/checkout/${data._id}`);
   } catch (error) {
     console.log(error);
-    alert("Error al realizar la compra");
+
+    Toastify({
+      text: "No se pudo realizar la compra",
+      duration: 3000,
+      destination: "https://github.com/apvarun/toastify-js",
+      newWindow: true,
+      gravity: "top", // `top` or `bottom`
+      position: "right", // `left`, `center` or `right`
+      stopOnFocus: true, // Prevents dismissing of toast on hover
+      style: {
+        background:
+          "linear-gradient(90deg, rgba(0,0,0,1) 0%, rgba(121,9,9,1) 36%, rgba(255,16,0,1) 100%)",
+      },
+      onClick: function () {}, // Callback after click
+    }).showToast();
   }
 };
